moved application class to app.py
